Bug 1464400 - Keep track of CSS generics when resolving to actual font families and faces, and expose as a new CSSGeneric attribute on InspectorFontFace. r=jwatt

+++ gfx/thebes/gfxFontFamilyList.h
@@ -23,7 +23,7 @@ namespace mozilla {
  * between unquoted and quoted names for serializaiton
  */ 
 
-enum FontFamilyType : uint32_t {
+enum FontFamilyType : uint8_t {
   eFamily_none = 0,  // used when finding generics
 
   // explicitly named font family (e.g. Helvetica)
